feat: small refactor
