+++ apps/client/src/features/app-settings/panel/project-panel/ProjectList.tsx
@@ -1,13 +1,6 @@
-import { IconButton, Input, Menu, MenuButton, MenuItem, MenuList } from '@chakra-ui/react';
-import { IoEllipsisHorizontal } from '@react-icons/all-files/io5/IoEllipsisHorizontal';
-
 import { useProjectList } from '../../../../common/hooks-query/useProjectList';
 import * as Panel from '../PanelUtils';
-
-import style from './ProjectPanel.module.scss';
-import { loadProject, renameProject } from '../../../../common/api/ontimeApi';
-import { useRef, useState } from 'react';
-import { IoSaveOutline } from '@react-icons/all-files/io5/IoSaveOutline';
+import ProjectListItem from './ProjectListItem';
 
 export default function ProjectList() {
   const { data, refetch } = useProjectList();
@@ -18,22 +11,9 @@ export default function ProjectList() {
   const projectFiles = [...files];
   const current = projectFiles.splice(currentlyLoadedIndex, 1)[0];
 
-  // TODO: Improve this
-  const [editing, setEditing] = useState<string | null>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-
   const handleRefetch = () => {
     refetch();
-  };
-
-  const handleToggleRename = (filename: string) => {
-    setEditing((prev) => (prev === filename ? null : filename));
-  };
-
-  const handleSubmitRename = () => {
-    renameProject(editing!, inputRef.current!.value);
-  };
-
+  }
 
   return (
     <Panel.Table>
@@ -47,102 +27,14 @@ export default function ProjectList() {
       </thead>
       <tbody>
         {current && (
-          <tr className={style.current}>
-            <td>{current.filename}</td>
-            <td>{new Date(current.createdAt).toLocaleString()}</td>
-            <td>{new Date(current.updatedAt).toLocaleString()}</td>
-            <td className={style.actionButton}>
-              <ActionMenu filename={current.filename} />
-            </td>
-          </tr>
+          <ProjectListItem filename={current.filename} createdAt={current.createdAt} updatedAt={current.updatedAt} onRefetch={handleRefetch} />
         )}
-        {projectFiles.map((project) => {
-          const createdAt = new Date(project.createdAt).toLocaleString();
-          const updatedAt = new Date(project.updatedAt).toLocaleString();
-          return (
-            <tr key={project.filename}>
-              {project.filename === editing ? (
-                <div
-                  style={{
-                    display: 'flex',
-                    flexDirection: 'row',
-                    alignItems: 'center',
-                    justifyContent: 'space-between',
-                    height: '100%',
-                  }}
-                >
-                  <Input
-                    size='sm'
-                    ref={inputRef}
-                    // data-testid='delay-input'
-                    className={style.inputField}
-                    type='text'
-                    variant='ontime-filled'
-                    // onFocus={handleFocus}
-                    // onChange={(event) => setValue(event.target.value)}
-                    // onBlur={(event) => validateAndSubmit(event.target.value)}
-                    // onKeyDown={onKeyDownHandler}
-                    // value={project.filename}
-                    defaultValue={project.filename}
-                />
-                  <IconButton
-                    size='sm'
-                    icon={<IoSaveOutline />}
-                    aria-label='Save duplicate project name'
-                    variant={'ontime-filled'}
-                    onClick={handleSubmitRename}
-                  />
-                </div>
-              ) : (
-                <td>{project.filename}</td>
-              )}
-              <td>{createdAt}</td>
-              <td>{updatedAt}</td>
-              <td className={style.actionButton}>
-                <ActionMenu filename={project.filename} onAction={handleRefetch} onRename={handleToggleRename} />
-              </td>
-            </tr>
-          );
-        })}
+        {
+          projectFiles.map((project) => (
+            <ProjectListItem key={project.filename} filename={project.filename} createdAt={project.createdAt} updatedAt={project.updatedAt} onRefetch={handleRefetch} />
+          ))
+        }
       </tbody>
     </Panel.Table>
-  );
-}
-
-function ActionMenu({
-  filename,
-  onAction,
-  onRename,
-}: {
-  filename: string;
-  onAction?: () => void;
-  onRename?: (filename: string) => void;
-}) {
-  const handleLoad = async () => {
-    await loadProject(filename);
-    onAction?.();
-    // TODO: Add a toast or something here
-  };
-
-  const handleRename = () => {
-    onRename?.(filename);
-  };
-
-  return (
-    <Menu variant='ontime-on-dark' size='sm'>
-      <MenuButton
-        as={IconButton}
-        aria-label='Options'
-        icon={<IoEllipsisHorizontal />}
-        variant='ontime-ghosted'
-        size='sm'
-      />
-      <MenuList>
-        <MenuItem onClick={handleLoad}>Load</MenuItem>
-        <MenuItem onClick={handleRename}>Rename</MenuItem>
-        <MenuItem>Duplicate</MenuItem>
-        <MenuItem>Delete</MenuItem>
-      </MenuList>
-    </Menu>
   );
 }
